FET: Extraido componente de podium para a pasta features
